updated code in movie service class

+++ SpringMovie/src/main/java/com/example/springmovie/service/MovieServiceImpl.java
@@ -1,67 +1,75 @@
-package com.example.springmovie.service.impl;
+package com.example.springmovie.service;
 
 import com.example.springmovie.enums.ReleaseStatus;
+import com.example.springmovie.exception.NotFoundException;
 import com.example.springmovie.model.Genre;
 import com.example.springmovie.model.Movie;
 import com.example.springmovie.repositories.MovieRepository;
-import com.example.springmovie.service.MovieService;
+import com.example.springmovie.service.interfaces.MovieService;
 import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
-import java.util.NoSuchElementException;
+import java.util.Optional;
 
 @RequiredArgsConstructor
-@Slf4j
-
 @Service
 public class MovieServiceImpl implements MovieService {
 
     private final MovieRepository movieRepository;
 
     @Override
-    public List<Movie> saveAll() {
-        return movieRepository.findAll();
+    public Page<Movie> findAllMovies(Pageable pageable) {
+        return movieRepository.findAll(pageable);
     }
 
     @Override
-    public Movie findById(Long id) {
-        return movieRepository.findById(id).orElseThrow(() -> new NoSuchElementException("Cannot find movie with id - " + id));
+    public Movie findMovieById(Long id) {
+        return movieRepository.findById(id)
+                .orElseThrow(() -> new NotFoundException(id));
     }
 
     @Override
-    public Movie save(Movie movie) {
+    public Movie saveMovie(Movie movie) {
         return movieRepository.save(movie);
     }
 
     @Override
     public Movie updateMovie(Movie movie) {
+        Optional<Movie> existingMovie = movieRepository.findById(movie.getId());
+        if (existingMovie.isEmpty()) {
+            throw new NotFoundException(movie.getId());
+        }
         return movieRepository.save(movie);
     }
 
     @Override
-    public void deleteById(Long id) {
+    public void deleteMovieById(Long id) {
+        if (!movieRepository.existsById(id)) {
+            throw new NotFoundException(id);
+        }
         movieRepository.deleteById(id);
     }
 
     @Override
-    public List<Movie> findByGenreId(Long genreId) {
+    public List<Movie> findMovieByGenreId(Long genreId) {
         return movieRepository.findMoviesByGenreId(genreId);
     }
 
     @Override
-    public List<Movie> findByGenreName(String genreName) {
+    public List<Movie> findMovieByGenreName(String genreName) {
         return movieRepository.findMoviesByGenreName(genreName);
     }
 
     @Override
-    public List<Movie> findByReleaseStatus(ReleaseStatus status) {
+    public List<Movie> findMovieByReleaseStatus(ReleaseStatus status) {
         return movieRepository.findMovieByReleaseStatus(status);
     }
 
     @Override
-    public List<Movie> findByYear(int year) {
+    public List<Movie> findMovieByYear(int year) {
         return movieRepository.findMoviesByYear(year);
     }
